refactor(web): merge Column into Columns component
