+++ web/src/components/Render.tsx
@@ -1,7 +1,6 @@
 import type React from 'react';
 import Button from '@/components/viavai/Button';
-import Column from '@/components/viavai/Column';
-import Columns from '@/components/viavai/Columns';
+import Columns, { Column } from '@/components/viavai/Columns';
 import Dialog from '@/components/viavai/Dialog';
 import Hero from '@/components/viavai/Hero';
 import Layout from '@/components/viavai/Layout';
